Move email to controllers folder and set sender name
